Auth

+++ notes/views.py
@@ -6,6 +6,7 @@ from rest_framework.response import Response
 
 from .models import Note
 from .serializers import NoteSerializer
+from constants import AuthConstants
 
 
 @api_view(['GET'])
@@ -22,53 +23,85 @@ def note_api_overview(request):
 
 @api_view(['GET'])
 def note_list(request):
-    notes = Note.objects.all().order_by('-timestamp')
-    serializer = NoteSerializer(notes, many=True)
-    return Response(serializer.data)
+    if request.user.is_authenticated:
+        notes = Note.objects.filter(user=request.user)
+        serializer = NoteSerializer(notes, many=True)
+        return Response(serializer.data)
+
+    err_res = {
+        'code': AuthConstants.TOKEN_MISSING,
+        'detail': 'No token was provided.'
+    }
+    return Response(err_res, status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET'])
 def note_detail(request, pk):
-    try:
-        note = Note.objects.get(pk=pk)
-    except ObjectDoesNotExist:
-        return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
-    serializer = NoteSerializer(note, many=False)
-    return Response(serializer.data)
+    if request.user.is_authenticated:
+        try:
+            note = Note.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
+
+        if note.user.id != request.user.id:
+            return Response('You can\'t access this note.', status=status.HTTP_401_UNAUTHORIZED)
+
+        serializer = NoteSerializer(note, many=False)
+        return Response(serializer.data)
+
+    return Response('You must be logged in.', status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['POST'])
 def note_create(request):
-    serializer = NoteSerializer(data=request.data)
+    if request.user.is_authenticated:
+        request.data['user'] = request.user.id
+        serializer = NoteSerializer(data=request.data)
 
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    return Response('You must be logged in.', status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['PUT'])
 def note_update(request, pk):
-    try:
-        note = Note.objects.get(pk=pk)
-    except ObjectDoesNotExist:
-        return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
+    if request.user.is_authenticated:
+        try:
+            note = Note.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
 
-    serializer = NoteSerializer(instance=note, data=request.data)
+        if note.user.id != request.user.id:
+            return Response('You can\'t access this note.', status=status.HTTP_401_UNAUTHORIZED)
 
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
+        request.data['user'] = request.user.id
+        serializer = NoteSerializer(instance=note, data=request.data)
 
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    return Response('You must be logged in.', status=status.HTTP_401_UNAUTHORIZED)
+
 
 @api_view(['DELETE'])
 def note_delete(request, pk):
-    try:
-        note = Note.objects.get(pk=pk)
-    except ObjectDoesNotExist:
-        return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
+    if request.user.is_authenticated:
+        try:
+            note = Note.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response('Invalid note ID.', status=status.HTTP_404_NOT_FOUND)
 
-    note.delete()
-    return Response('Note deleted.')
+        if note.user.id != request.user.id:
+            return Response('You can\'t access this note.', status=status.HTTP_401_UNAUTHORIZED)
+
+        note.delete()
+        return Response('Note deleted.')
+
+    return Response('You must be logged in.', status=status.HTTP_401_UNAUTHORIZED)
